refactor: Extract sails helpers to lib * getNumberOfReviews * getNumberOfStars * getAvgNumberOfStars
