+++ lib/getNumberOfReviews.js
@@ -1,0 +1,4 @@
+module.exports = (ratings, placeId) => {
+  const numberOfReviews = ratings.filter(rating => rating.placeId === placeId).length;
+  return numberOfReviews;
+};
